split hooks and components into different files, fix find deposit logic, add styles and progress for tx dialogs
