+++ libs/ui-toolkit/src/components/dialog/dialog.tsx
@@ -1,8 +1,7 @@
 import * as DialogPrimitives from '@radix-ui/react-dialog';
 import classNames from 'classnames';
 import type { ReactNode } from 'react';
-import type { Intent } from '../../utils/intent';
-import { getIntentShadow } from '../../utils/intent';
+import { Intent } from '../../utils/intent';
 import { Icon } from '../icon';
 
 interface DialogProps {
@@ -25,7 +24,15 @@ export function Dialog({
     'fixed w-[520px] px-28 py-24 top-[50%] left-[50%] translate-x-[-50%] translate-y-[-50%]',
     // Need to apply background and text colors again as content is rendered in a portal
     'dark:bg-black dark:text-white-95 bg-white text-black-95',
-    getIntentShadow(intent)
+    'shadow-callout',
+    {
+      'shadow-intent-danger': intent === Intent.Danger,
+      'shadow-intent-warning': intent === Intent.Warning,
+      'shadow-intent-prompt': intent === Intent.Prompt,
+      'shadow-black dark:shadow-white': intent === Intent.Progress,
+      'shadow-intent-success': intent === Intent.Success,
+      'shadow-intent-help': intent === Intent.Help,
+    }
   );
   return (
     <DialogPrimitives.Root open={open} onOpenChange={(x) => onChange(x)}>
